+++ tests/DummyPlugin/src/Modules/MethodCallModule.php
@@ -1,0 +1,13 @@
+<?php
+
+namespace Bojaghi\Continy\Tests\DummyPlugin\Modules;
+
+final class MethodCallModule
+{
+    public static bool $called = false;
+
+    public function testMethod(): void
+    {
+        self::$called = true;
+    }
+}